functions/requirements.txt
firebase-admin>=5.0.0
scikit-learn==1.2.2
numpy==1.23.5
google-cloud-firestore==2.10.0
functions-framework==3.1.0
pandas==1.5.3
